requirements.txt
discord.py
Flask
requests
python-dotenv
geopandas
matplotlib
shapely
plotly
kaleido